# docs/swagger.yaml
basePath: /
definitions:
  main.AirportResponse:
    properties:
      country:
        type: string
      iata:
        type: string
      icao:
        type: string
      lat:
        type: number
      lng:
        type: number
      municipality:
        type: string
      name:
        type: string
      radius:
        type: number
      region:
        type: string
      type:
        type: string
    type: object
  main.AirportSearchResponse:
    properties:
      count:
        type: integer
      query:
        type: string
      results:
        items:
          $ref: '#/definitions/main.AirportResponse'
        type: array
    type: object
  main.BoundingBoxResponse:
    properties:
      lamax:
        type: number
      lamin:
        type: number
      lomax:
        type: number
      lomin:
        type: number
    type: object
  main.CacheStats:
    properties:
      entries:
        type: integer
      ttl:
        type: string
    type: object
  main.CacheStatsResponse:
    properties:
      flightCache:
        $ref: '#/definitions/main.CacheStats'
    type: object
  main.ErrorResponse:
    properties:
      error:
        type: string
      example:
        type: string
      hint:
        type: string
    type: object
  main.Flight:
    properties:
      altitude:
        description: Barometric altitude in meters
        type: number
      callsign:
        description: Flight number (pointer = nullable)
        type: string
      icao24:
        description: Unique aircraft identifier
        type: string
      last_contact:
        description: Unix timestamp of last contact
        type: integer
      latitude:
        description: WGS-84 latitude
        type: number
      longitude:
        description: WGS-84 longitude
        type: number
      on_ground:
        description: Is the aircraft on ground?
        type: boolean
      origin_country:
        description: Country of registration
        type: string
      velocity:
        description: Ground speed in m/s
        type: number
    type: object
  main.FlightsByAirportResponse:
    properties:
      airport:
        $ref: '#/definitions/main.AirportResponse'
      bbox:
        $ref: '#/definitions/main.BoundingBoxResponse'
      count:
        type: integer
      flights:
        items:
          $ref: '#/definitions/main.Flight'
        type: array
      time:
        type: integer
    type: object
  main.FlightsResponse:
    properties:
      count:
        type: integer
      flights:
        items:
          $ref: '#/definitions/main.Flight'
        type: array
      time:
        type: integer
    type: object
  main.PingResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
  description: Flight tracking API with airport search, caching, and rate limiting.
  title: OpenSky Go API
  version: "1.0"
paths:
  /airports:
    get:
      description: Search airports by ICAO, IATA, airport name, or city
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - description: Max results (1-50, default 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.AirportSearchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Search airports
      tags:
      - airports
  /airports/{icao}:
    get:
      description: Get airport details by 4-letter ICAO code
      parameters:
      - description: Airport ICAO code
        in: path
        name: icao
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.AirportResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Get airport by ICAO
      tags:
      - airports
  /cache/stats:
    get:
      description: Returns current in-memory cache statistics
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.CacheStatsResponse'
      summary: Cache stats
      tags:
      - system
  /flights/airport/{icao}:
    get:
      description: Get flights near an airport using a radius-based bounding box
      parameters:
      - description: Airport ICAO code
        in: path
        name: icao
        required: true
        type: string
      - description: Radius in degrees (0-5, default 0.3)
        in: query
        name: radius
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.FlightsByAirportResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: List flights near airport
      tags:
      - flights
  /flights/area:
    get:
      description: Get flights within a geographic bounding box
      parameters:
      - description: Minimum latitude
        in: query
        name: lamin
        required: true
        type: number
      - description: Maximum latitude
        in: query
        name: lamax
        required: true
        type: number
      - description: Minimum longitude
        in: query
        name: lomin
        required: true
        type: number
      - description: Maximum longitude
        in: query
        name: lomax
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.FlightsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: List flights in area
      tags:
      - flights
  /ping:
    get:
      description: Basic health check endpoint
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.PingResponse'
      summary: Health check
      tags:
      - system
swagger: "2.0"